fixed upgrade

+++ migration/migration.go
@@ -320,44 +320,7 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		},
 	}
 
-	tableVersions := []model.TableVersions{
-		model.TableVersions{
-			NameTable:    "versions",
-			VersionTable: jsonschema.TableVersion["versions"],
-		},
-		model.TableVersions{
-			NameTable:    "applications",
-			VersionTable: jsonschema.TableVersion["applications"],
-		},
-		model.TableVersions{
-			NameTable:    "agent_location_session",
-			VersionTable: jsonschema.TableVersion["agent_location_session"],
-		},
-		model.TableVersions{
-			NameTable:    "alias",
-			VersionTable: jsonschema.TableVersion["alias"],
-		},
-		model.TableVersions{
-			NameTable:    "global_settings",
-			VersionTable: jsonschema.TableVersion["global_settings"],
-		},
-		model.TableVersions{
-			NameTable:    "hepsub_mapping_schema",
-			VersionTable: jsonschema.TableVersion["hepsub_mapping_schema"],
-		},
-		model.TableVersions{
-			NameTable:    "mapping_schema",
-			VersionTable: jsonschema.TableVersion["mapping_schema"],
-		},
-		model.TableVersions{
-			NameTable:    "users",
-			VersionTable: jsonschema.TableVersion["users"],
-		},
-		model.TableVersions{
-			NameTable:    "user_settings",
-			VersionTable: jsonschema.TableVersion["user_settings"],
-		},
-	}
+	tableVersions := []model.TableVersions{}
 
 	mappingSchema := []model.TableMappingSchema{
 		model.TableMappingSchema{
@@ -506,6 +469,10 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range usersData {
 			configDBSession.Save(&el)
 		}
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "users",
+			VersionTable: jsonschema.TableVersion["users"],
+		})
 	}
 
 	forceIt = force
@@ -520,6 +487,11 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range globalSettingData {
 			configDBSession.Save(&el)
 		}
+
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "global_settings",
+			VersionTable: jsonschema.TableVersion["global_settings"],
+		})
 	}
 
 	forceIt = force
@@ -534,6 +506,10 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range agentLocationSession {
 			configDBSession.Save(&el)
 		}
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "agent_location_session",
+			VersionTable: jsonschema.TableVersion["agent_location_session"],
+		})
 	}
 
 	forceIt = force
@@ -548,6 +524,11 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range hepsubSchema {
 			configDBSession.Save(&el)
 		}
+
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "hepsub_mapping_schema",
+			VersionTable: jsonschema.TableVersion["hepsub_mapping_schema"],
+		})
 	}
 
 	forceIt = force
@@ -562,6 +543,11 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range dashboardUsers {
 			configDBSession.Save(&el)
 		}
+
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "user_settings",
+			VersionTable: jsonschema.TableVersion["user_settings"],
+		})
 	}
 
 	forceIt = force
@@ -576,13 +562,26 @@ func PopulateHomerConfigTables(configDBSession *gorm.DB, homerDBconfig string, f
 		for _, el := range mappingSchema {
 			configDBSession.Save(&el)
 		}
+
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "mapping_schema",
+			VersionTable: jsonschema.TableVersion["mapping_schema"],
+		})
 	}
 
-	/* tableVersions data */
-	heputils.Colorize(heputils.ColorRed, "reinstalling versions")
-	configDBSession.Exec("TRUNCATE TABLE versions")
-	for _, el := range tableVersions {
-		configDBSession.Save(&el)
+	if len(tableVersions) > 0 {
+
+		tableVersions = append(tableVersions, model.TableVersions{
+			NameTable:    "versions",
+			VersionTable: jsonschema.TableVersion["versions"],
+		})
+
+		/* tableVersions data */
+		heputils.Colorize(heputils.ColorRed, "reinstalling versions")
+		configDBSession.Exec("TRUNCATE TABLE versions")
+		for _, el := range tableVersions {
+			configDBSession.Save(&el)
+		}
 	}
 
 	heputils.Colorize(heputils.ColorYellow, "\r\nDONE")
